Extract `Label`'s text layout cache into a shared `TextLayoutCache`.
Pure refactor: `TextLayout` and the caching logic introduced for `Label`
move verbatim to a new crate-private `widgets::text_layout_cache` module,
parameterized on the text and styles instead of reading `Label`'s fields.
This makes the cache usable by `TextArea` measurement in the next commit.

No behavior change; `Label`'s render snapshot tests are unchanged.
