+++ masonry/src/widgets/mod.rs
@@ -42,6 +42,7 @@ mod svg;
 mod switch;
 mod text_area;
 mod text_input;
+mod text_layout_cache;
 mod variable_label;
 mod virtual_scroll;
 mod zstack;
